[DDW-821] Restructures test setup helpers and types
