--- features/tests/e2e/setup/webdriver.js
@@ -1,8 +1,0 @@
-import { Before } from 'cucumber';
-
-Before(function() {
-  this.waitAndClick = async (selector, ...waitArgs) => {
-    await this.client.waitForVisible(selector, ...waitArgs);
-    return this.client.click(selector);
-  };
-});
