--- features/tests/unit/setup/context.js
@@ -1,5 +1,0 @@
-import { Before } from 'cucumber';
-
-Before(function() {
-  this.context = {};
-});
