--- features/tests/unit/setup/parameter-types.js
@@ -1,8 +1,0 @@
-import { defineParameterType } from 'cucumber';
-
-// Add {bool} parameter type
-defineParameterType({
-  name: 'bool',
-  regexp: /true|false/,
-  transformer: b => b === 'true',
-});
